Create a factor for a CFTime instance

Method for base::cut() applied to CFTime objects.

Usage

# S3 method for class 'CFTime'
cut(x, breaks, ...)

Arguments

x

An instance of CFTime.

breaks

A character string of a factor period (see CFfactor() for a description), or a character vector of timestamps that conform to the calendar of x, with a length of at least 2. Timestamps must be given in ISO8601 format, e.g. "2024-04-10 21:31:43".

...

Ignored.

Value

A factor with levels according to the breaks argument, with attributes 'period', 'era' and 'CFTime'. When breaks is a factor period, attribute 'period' has that value, otherwise it is '"day"'. When breaks is a character vector of timestamps, attribute 'CFTime' holds an instance of CFTime that has the same definition as x, but with (ordered) offsets generated from the breaks. Attribute 'era' is always -1.

Details

When breaks is one of "year", "season", "quarter", "month", "dekad", "day" a factor is generated like by CFfactor().

When breaks is a vector of character timestamps a factor is produced with a level for every interval between timestamps. The last timestamp, therefore, is only used to close the interval started by the pen-ultimate timestamp - use a distant timestamp (e.g. range(x)[2]) to ensure that all offsets to the end of the CFTime time series are included, if so desired. The last timestamp will become the upper bound in the CFTime instance that is returned as an attribute to this function so a sensible value for the last timestamp is advisable.

This method works similar to base::cut.POSIXt() but there are some differences in the arguments: for breaks the set of options is different and no preceding integer is allowed, labels are always assigned using values of breaks, and the interval is always left-closed.

See also

Examples

x <- CFtime("days since 2021-01-01", "365_day", 0:729)
breaks <- c("2022-02-01", "2021-12-01", "2023-01-01")
cut(x, breaks)
#>   [1] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>   [7] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [13] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [19] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [25] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [31] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [37] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [43] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [49] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [55] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [61] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [67] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [73] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [79] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [85] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [91] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#>  [97] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [103] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [109] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [115] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [121] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [127] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [133] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [139] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [145] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [151] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [157] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [163] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [169] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [175] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [181] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [187] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [193] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [199] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [205] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [211] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [217] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [223] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [229] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [235] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [241] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [247] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [253] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [259] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [265] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [271] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [277] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [283] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [289] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [295] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [301] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [307] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [313] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [319] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [325] <NA>       <NA>       <NA>       <NA>       <NA>       <NA>      
#> [331] <NA>       <NA>       <NA>       <NA>       2021-12-01 2021-12-01
#> [337] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [343] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [349] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [355] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [361] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [367] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [373] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [379] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [385] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [391] 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01 2021-12-01
#> [397] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [403] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [409] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [415] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [421] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [427] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [433] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [439] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [445] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [451] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [457] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [463] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [469] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [475] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [481] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [487] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [493] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [499] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [505] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [511] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [517] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [523] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [529] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [535] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [541] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [547] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [553] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [559] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [565] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [571] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [577] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [583] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [589] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [595] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [601] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [607] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [613] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [619] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [625] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [631] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [637] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [643] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [649] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [655] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [661] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [667] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [673] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [679] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [685] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [691] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [697] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [703] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [709] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [715] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [721] 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> [727] 2022-02-01 2022-02-01 2022-02-01 2022-02-01
#> attr(,"period")
#> [1] day
#> attr(,"era")
#> [1] -1
#> attr(,"CFTime")
#> CF calendar:
#>   Origin  : 2021-01-01T00:00:00
#>   Units   : days
#>   Type    : 365_day
#> Time series:
#>   Elements: [2022-01-01 .. 2022-07-18] (average of 198.000000 days between 2 elements)
#>   Bounds  : set
#> Levels: 2021-12-01 2022-02-01