Deprecated functions

These functions are deprecated and should no longer be used in new code. The below table gives the replacement function to use instead. The function arguments of the replacement function are the same as those of the deprecated function if no arguments are given in the table.

Deprecated function	Replacement function	Secheduled deletion

slab()	slice()	October 2025

The below functions have been permanently deleted and can no longer be called:

Deprecated function	Replacement function

CFcomplete()	is_complete()

CFmonth_days()	month_days()

CFparse()	parse_timestamps()

CFrange()	range()

CFsubset()	slice()

slab()	slice()

CFtimestamp()	as_timestamp()

Usage

slab(x, extremes, rightmost.closed = FALSE)

Arguments

x, extremes, rightmost.closed

See replacement functions.

Value

See replacement functions.