Return the number of days in a month given a certain CF calendar

Given a vector of dates as strings in ISO 8601 or UDUNITS format and a CFTime object, this function will return a vector of the same length as the dates, indicating the number of days in the month according to the calendar specification. If no vector of days is supplied, the function will return an integer vector of length 12 with the number of days for each month of the calendar (disregarding the leap day for standard and julian calendars).

Usage

month_days(t, x = NULL)

Arguments

t

The CFTime instance to use.

x

character. An optional vector of dates as strings with format YYYY-MM-DD. Any time part will be silently ingested.

Value

A vector indicating the number of days in each month for the vector of dates supplied as argument x. Invalidly specified dates will result in an NA value. If no dates are supplied, the number of days per month for the calendar as a vector of length 12.

See also

Examples

dates <- c("2021-11-27", "2021-12-10", "2022-01-14", "2022-02-18")
t <- CFtime("days since 1850-01-01", "standard")
month_days(t, dates)
#> [1] 30 31 31 28

t <- CFtime("days since 1850-01-01", "360_day")
month_days(t, dates)
#> [1] 30 30 30 30

t <- CFtime("days since 1850-01-01", "all_leap")
month_days(t, dates)
#> [1] 30 31 31 29

month_days(t)
#>  [1] 31 29 31 30 31 30 31 31 30 31 30 31